test(email): move importance pagination tests into their own file; keep fixture/test comments behavioral
