+++ rust/cloud-storage/email/src/outbound/email_pg_repo/test/mod.rs
@@ -1,6 +1,7 @@
 mod crm_scope_dynamic_query;
 mod draft;
 mod dynamic_query;
+mod importance_pagination;
 mod labels;
 mod link;
 mod message;
